Move content of mime to main module
